+++ tests/cloudpickle_test.py
@@ -1236,7 +1236,9 @@ class CloudPickleTest(unittest.TestCase):
 
         self.assertEqual(f2.__doc__, f.__doc__)
 
-    @unittest.skipIf(sys.version_info[0] < 3, "This syntax won't work on py2.")
+    @unittest.skipIf(sys.version_info < (3, 7),
+                     """This syntax won't work on py2 and pickling annotations
+                     isn't supported for py36 and below.""")
     def test_wraps_preserves_function_annotations(self):
         from functools import wraps